requirements.txt
django
djangorestframework
psycopg2
pillow
django-cors-headers
djangorestframework-simplejwt
drf-yasg
